refactor(do): extract shared FTS and aggregate-tally primitives
Move the dialect-agnostic helpers out of ctx-db.ts into two reusable
modules so @cirrus/d1 can import them instead of carrying byte-identical
copies (the drift risk flagged in review):

- search-text.ts: tokenizeSearch, buildFtsMatch, stringifySearchText,
  scoreDocument, ftsTableName
- aggregate-tally.ts: coerceAggregateNumber, AggregateTally,
  foldAggregateTally, readAggregateValue, aggregateTableName,
  encodeAggregateKey

Re-exported from the package root. Pure move, bodies unchanged.
